chore: refactor alert polyfill
