+++ src/hooks/useTimestamp.ts
@@ -1,11 +1,16 @@
 import { useEffect } from 'react'
 
+import { alert } from '~utils'
+
 export const useTimestamp = (): void => {
   useEffect(() => {
     fetch('/api/timestamp')
       .then((response) => response.json())
       .then((data) => {
-        alert(`This is just an example response form miragejs \n\n ${JSON.stringify(data)}`)
+        alert(
+          'Warning',
+          `This is just an example response form miragejs \n\n ${JSON.stringify(data)}`
+        )
       })
   }, [])
 }
